// resources/js/types/diser.ts
export type Diser = {
    id: number;
    name: string;
    rsc_re: string;
    fb_name: string;
    rate: number;
    sales: number;
    total: number;
    others_1: string;
    hold_stop_allow: string;
    gcash_number: string;
    gcash_name: string;
    sv_only: string;
    company_sv: string;
    others_2: string;
    others_3: string;
};